Added argparse commandline arguments
An older version of htis application was a commandline utility that
accepted arguements for the target directory, whether to unzip the
archives first, verbosity, and whether to append or rebuild the target
database. Those arguments have been re-created to the main.py script can
be used as a command line utility again.